FILE: moves getValidCombos
