+++ libs/spirit-islander/util/src/lib/app/get-valid-combos.ts
@@ -1,5 +1,5 @@
 import type { DifficultyOption } from '../game/difficulty';
-import type { Config } from '../app/config.interface';
+import type { Config } from './config.interface';
 import { MAPS, Map, MapName } from '../game/maps';
 import { SCENARIOS, Scenario, ScenarioName } from '../game/scenarios';
 import {
@@ -7,7 +7,7 @@ import {
   AdversaryLevel,
   AdversaryLevelName,
 } from '../game/adversaries';
-import { ComboAnalyzer } from '../app/combo-analyzer';
+import { ComboAnalyzer } from './combo-analyzer';
 import { getDifficulty } from '../game/get-difficulty';
 
 const comboAnalyzer = new ComboAnalyzer<
